test(downloads): split module browser coverage
Why:
- Continue the test architecture migration by removing package-owned module-browser coverage from the app-host unit bundle.
- Keep Android Downloads and metadata parity contracts enforced in the owning package lanes.

What Changed:
- Moved Downloads filtering, failed-source cache merge, and startup-default installable-row tests into BibleUITests.
- Added SwordKit metadata tests for Android recommendation tokens and catalog InstallSize byte preservation.
- Kept recommended-document refresh failure coverage in SwordManagerTests and removed the duplicate app-host copy.
- Updated docs/test-architecture-migration-plan.md with the AppAndReader module-browser slice status.

Validation:
- python3 scripts/check_repo_standards.py docblocks --all-files
- python3 scripts/check_repo_standards.py source-guards
- git diff --check
- xcodebuild test -project AndBible.xcodeproj -scheme BibleUITests -destination 'platform=iOS Simulator,name=iPhone 17' -derivedDataPath /tmp/andbible-module-browser-bibleui-dd CODE_SIGNING_ALLOWED=NO -only-testing:BibleUITests/ModuleBrowserDownloadsTests
- xcodebuild test -project AndBible.xcodeproj -scheme SwordKitTests -destination 'platform=iOS Simulator,name=iPhone 17' -derivedDataPath /tmp/andbible-module-browser-swordkit-dd CODE_SIGNING_ALLOWED=NO -only-testing:SwordKitTests/ModuleDownloadMetadataTests -only-testing:SwordKitTests/SwordManagerTests/testRecommendedDocumentRefreshPreservesCachedMetadataAfterFailures
- xcodebuild test -project AndBible.xcodeproj -scheme AndBible -destination 'platform=iOS Simulator,name=iPhone 17' -derivedDataPath /tmp/andbible-module-browser-app-dd CODE_SIGNING_ALLOWED=NO -only-testing:AndBibleTests/AndBibleTests/testBibleReaderModulePickerBuildsForBibleCategory
- GitNexus detect_changes scope=staged returned no mapped production-flow changes.
- Adversarial subagent review found no blocking issues.

Impact:
- Reduces app-host test surface while preserving Android module-browser and metadata parity checks.
- Production behavior is unchanged; risk is limited to test placement and coverage mapping.
