+++ AndBibleTests/AndBibleTests+AppAndReader.swift
@@ -16,353 +16,6 @@ import struct SwiftUI.Color
 #endif
 
 extension AndBibleTests {
-    func testModuleBrowserFiltersAndSortsAndroidDownloadRows() {
-        let modules = [
-            RemoteModuleInfo(
-                name: "KJV",
-                description: "King James Version",
-                category: .bible,
-                language: "en",
-                sourceName: "CrossWire",
-                version: "1.0",
-                installSizeBytes: 1_260_000
-            ),
-            RemoteModuleInfo(
-                name: "WEB",
-                description: "World English Bible",
-                category: .bible,
-                language: "en",
-                sourceName: "CrossWire",
-                version: "2.0"
-            ),
-            RemoteModuleInfo(
-                name: "REC",
-                description: "Recommended Bible",
-                category: .bible,
-                language: "en",
-                sourceName: "CrossWire",
-                version: "1.0"
-            ),
-            RemoteModuleInfo(
-                name: "WARN",
-                description: "Known warning module",
-                category: .bible,
-                language: "en",
-                sourceName: "CrossWire",
-                version: "1.0"
-            ),
-            RemoteModuleInfo(
-                name: "HIDE",
-                description: "Hidden module",
-                category: .bible,
-                language: "en",
-                sourceName: "CrossWire",
-                version: "1.0"
-            ),
-            RemoteModuleInfo(
-                name: "MHC",
-                description: "Matthew Henry",
-                category: .commentary,
-                language: "en",
-                sourceName: "CrossWire",
-                version: "1.0"
-            ),
-            RemoteModuleInfo(
-                name: "ADDON",
-                description: "Add-on font pack",
-                category: .addon,
-                language: "zxx",
-                sourceName: "AndBible",
-                version: "1.0"
-            ),
-            RemoteModuleInfo(
-                name: "PSEUDO",
-                description: "Unavailable translation",
-                category: .bible,
-                language: "en",
-                sourceName: "Not Available",
-                availability: .unavailable,
-                unavailableReason: "Unavailable",
-                version: "0.0"
-            )
-        ]
-        let installed = [
-            ModuleInfo(name: "KJV", description: "King James Version", category: .bible, language: "en", version: "1.0"),
-            ModuleInfo(name: "WEB", description: "World English Bible", category: .bible, language: "en", version: "1.0")
-        ]
-        let recommended = ModuleDownloadConfiguration(
-            bibles: ["en": ["REC::CrossWire"]]
-        )
-        let bad = ModuleDownloadConfiguration(
-            bibles: ["en": ["WARN::CrossWire::1.0::W", "HIDE::CrossWire::1.0::H"]]
-        )
-
-        let filtered = ModuleBrowserView.filteredDownloadModules(
-            modules,
-            selectedCategory: nil,
-            selectedLanguage: "en",
-            searchText: "",
-            installedModules: installed,
-            downloadActivities: ["WARN": .inProgress(0.25)],
-            recommendedDocuments: recommended,
-            badDocuments: bad
-        )
-
-        XCTAssertEqual(filtered.map(\.name), ["WARN", "WEB", "KJV", "REC", "PSEUDO", "MHC"])
-        let addonFiltered = ModuleBrowserView.filteredDownloadModules(
-            modules,
-            selectedCategory: .addon,
-            selectedLanguage: "en",
-            searchText: "",
-            installedModules: installed,
-            downloadActivities: [:],
-            recommendedDocuments: recommended,
-            badDocuments: bad
-        )
-        XCTAssertEqual(addonFiltered.map(\.name), ["ADDON"])
-        XCTAssertEqual(
-            ModuleBrowserView.displayStatus(
-                for: modules[1],
-                installedModules: installed,
-                downloadActivities: [:]
-            ),
-            .updateAvailable
-        )
-        XCTAssertEqual(ModuleBrowserView.installSizeText(for: modules[0].installSizeBytes), "1.3 MB")
-        XCTAssertTrue(ModuleBrowserView.isRemoteVersionNewer(remoteVersion: "1.10", installedVersion: "1.9"))
-        XCTAssertFalse(ModuleBrowserView.isRemoteVersionNewer(remoteVersion: "1.0", installedVersion: "1.0"))
-        XCTAssertFalse(ModuleBrowserView.isRemoteVersionNewer(remoteVersion: "", installedVersion: "1.0"))
-        XCTAssertNil(ModuleBrowserView.installSizeText(for: 0))
-        XCTAssertNil(ModuleBrowserView.installSizeText(for: -1))
-    }
-
-    func testModuleBrowserMergesCachedCatalogRowsForFailedSources() {
-        let refreshedModules = [
-            RemoteModuleInfo(
-                name: "ASV",
-                description: "American Standard Version",
-                category: .bible,
-                language: "en",
-                sourceName: "AndBible"
-            ),
-            RemoteModuleInfo(
-                name: "KJV",
-                description: "King James Version",
-                category: .bible,
-                language: "en",
-                sourceName: "CrossWire"
-            )
-        ]
-        let cachedModules = [
-            RemoteModuleInfo(
-                name: "ASV",
-                description: "American Standard Version",
-                category: .bible,
-                language: "en",
-                sourceName: "AndBible"
-            ),
-            RemoteModuleInfo(
-                name: "KJV",
-                description: "King James Version",
-                category: .bible,
-                language: "en",
-                sourceName: "CrossWire"
-            ),
-            RemoteModuleInfo(
-                name: "MHC",
-                description: "Matthew Henry",
-                category: .commentary,
-                language: "en",
-                sourceName: "CrossWire"
-            ),
-            RemoteModuleInfo(
-                name: "WEB",
-                description: "World English Bible",
-                category: .bible,
-                language: "en",
-                sourceName: "CustomSource"
-            )
-        ]
-
-        let merged = ModuleBrowserView.modulesByAddingCachedCatalogsForFailedSources(
-            refreshedModules: refreshedModules,
-            cachedModules: cachedModules,
-            failedSourceNames: ["CrossWire"]
-        )
-
-        XCTAssertEqual(merged.map(\.id), ["AndBible:ASV", "CrossWire:KJV", "CrossWire:MHC"])
-    }
-
-    func testModuleBrowserStartupDefaultsRequireInstallableCatalogRows() {
-        let unavailableModules = [
-            RemoteModuleInfo(
-                name: "PSEUDO",
-                description: "Unavailable translation",
-                category: .bible,
-                language: "en",
-                sourceName: "Not Available",
-                availability: .unavailable,
-                unavailableReason: "Unavailable"
-            )
-        ]
-        let installableModules = [
-            RemoteModuleInfo(
-                name: "KJV",
-                description: "King James Version",
-                category: .bible,
-                language: "en",
-                sourceName: "CrossWire"
-            )
-        ]
-
-        XCTAssertFalse(ModuleBrowserView.startupDefaultCatalogHasInstallableRows([]))
-        XCTAssertFalse(ModuleBrowserView.startupDefaultCatalogHasInstallableRows(unavailableModules))
-        XCTAssertTrue(ModuleBrowserView.startupDefaultCatalogHasInstallableRows(installableModules))
-    }
-
-    func testDownloadConfigurationDecodesAndroidMetadataEntries() throws {
-        let data = """
-        {
-          "bibles": {"en": ["KJV::CrossWire", "ASV"]},
-          "commentaries": {},
-          "dictionaries": {},
-          "books": {},
-          "maps": {},
-          "addons": {"en": ["AddonFonts::AndBible"]}
-        }
-        """.data(using: .utf8)!
-        let recommended = try JSONDecoder().decode(ModuleDownloadConfiguration.self, from: data)
-        let bad = ModuleDownloadConfiguration(
-            bibles: ["en": ["KJV::CrossWire::2.3::W", "WEB::CrossWire::1.0::H"]],
-            addons: ["en": ["AddonFonts::AndBible::1.0::W"]]
-        )
-        let kjv = RemoteModuleInfo(
-            name: "KJV",
-            description: "King James Version",
-            category: .bible,
-            language: "en",
-            sourceName: "CrossWire",
-            version: "2.3"
-        )
-        let asv = RemoteModuleInfo(
-            name: "ASV",
-            description: "American Standard Version",
-            category: .bible,
-            language: "en",
-            sourceName: "Different",
-            version: "1.0"
-        )
-        let web = RemoteModuleInfo(
-            name: "WEB",
-            description: "World English Bible",
-            category: .bible,
-            language: "en",
-            sourceName: "CrossWire",
-            version: "1.0"
-        )
-        let addon = RemoteModuleInfo(
-            name: "AddonFonts",
-            description: "Add-on font pack",
-            category: .addon,
-            language: "zxx",
-            sourceName: "AndBible",
-            version: "1.0"
-        )
-
-        XCTAssertTrue(recommended.contains(kjv))
-        XCTAssertTrue(recommended.contains(asv))
-        XCTAssertTrue(recommended.contains(addon))
-        XCTAssertEqual(recommended.addons, ["en": ["AddonFonts::AndBible"]])
-        XCTAssertEqual(bad.badDocumentAction(for: kjv), .warn)
-        XCTAssertEqual(bad.badDocumentAction(for: addon), .warn)
-        XCTAssertEqual(bad.badDocumentAction(for: web), .hide)
-        XCTAssertEqual(bad.badDocumentAction(for: asv), .none)
-    }
-
-    func testRecommendedDocumentRefreshPreservesCachedMetadataAfterFailures() async throws {
-        let tempDir = FileManager.default.temporaryDirectory
-            .appendingPathComponent(UUID().uuidString, isDirectory: true)
-        let swordDir = tempDir.appendingPathComponent("sword", isDirectory: true)
-        try FileManager.default.createDirectory(at: tempDir, withIntermediateDirectories: true)
-        defer { try? FileManager.default.removeItem(at: tempDir) }
-
-        let validData = """
-        {
-          "bibles": {"en": ["KJV::CrossWire"]},
-          "commentaries": {},
-          "dictionaries": {},
-          "books": {},
-          "maps": {}
-        }
-        """.data(using: .utf8)!
-        let malformedData = Data("<html>temporary failure</html>".utf8)
-        var responses = [
-            (statusCode: 200, data: validData),
-            (statusCode: 200, data: malformedData),
-            (statusCode: 500, data: Data("temporary failure".utf8))
-        ]
-
-        MockURLProtocol.requestHandler = { request in
-            let responsePayload = responses.removeFirst()
-            let response = HTTPURLResponse(
-                url: request.url!,
-                statusCode: responsePayload.statusCode,
-                httpVersion: nil,
-                headerFields: nil
-            )!
-            return (response, responsePayload.data)
-        }
-
-        let repository = ModuleRepository(
-            basePath: tempDir.path,
-            swordPath: swordDir.path,
-            session: makeMockedURLSession()
-        )
-
-        let refreshedMetadata = try await repository.refreshRecommendedDocuments()
-        XCTAssertEqual(refreshedMetadata.bibles["en"], ["KJV::CrossWire"])
-        XCTAssertEqual(repository.loadCachedRecommendedDocuments()?.bibles["en"], ["KJV::CrossWire"])
-
-        do {
-            _ = try await repository.refreshRecommendedDocuments()
-            XCTFail("Expected malformed recommended-document metadata to fail decoding.")
-        } catch {
-            XCTAssertEqual(repository.loadCachedRecommendedDocuments()?.bibles["en"], ["KJV::CrossWire"])
-        }
-
-        do {
-            _ = try await repository.refreshRecommendedDocuments()
-            XCTFail("Expected non-200 recommended-document metadata to fail downloading.")
-        } catch {
-            XCTAssertEqual(repository.loadCachedRecommendedDocuments()?.bibles["en"], ["KJV::CrossWire"])
-        }
-    }
-
-    /**
-     Verifies iOS preserves Android's `InstallSize` units when rendering Downloads sizes.
-
-     Android reads SWORD `KEY_INSTALL_SIZE` directly as bytes and formats that value as megabytes.
-     iOS must not multiply it by 1024, because that inflates catalog sizes by three orders of
-     magnitude and makes the Downloads browser disagree with Android for the same repository row.
-     */
-    func testCatalogModulePreservesAndroidInstallSizeBytes() {
-        let module = CatalogModule(
-            name: "KJV",
-            description: "King James Version",
-            category: .bible,
-            language: "en",
-            modDrv: "zText",
-            dataPath: "modules/texts/ztext/kjv/",
-            confContent: "",
-            sourceName: "CrossWire",
-            version: "1.0",
-            size: "1260000"
-        )
-
-        XCTAssertEqual(module.remoteModuleInfo.installSizeBytes, 1_260_000)
-        XCTAssertEqual(ModuleBrowserView.installSizeText(for: module.remoteModuleInfo.installSizeBytes), "1.3 MB")
-    }
-
     func testBibleReaderModulePickerBuildsForBibleCategory() {
         let controller = BibleReaderController(bridge: BibleBridge(), initializesSword: false)
         let view = BibleReaderModulePicker(
