+++ docs/test-architecture-migration-plan.md
@@ -157,6 +157,7 @@ Blocker: `AndBibleTestSupport.swift` is a 2,233-line `extension AndBibleTests` e
 - `AndBibleTests+AppAndReader` SWORD coordinator slice has moved to `BibleUITests` because it exercises BibleUI reader SWORD setup, installed-module categorization, active-module fallback, book-list derivation, and display-option application without app delegate/bootstrap behavior.
 - `AndBibleTests+AppAndReader` module-switch planner slice has moved to `BibleUITests` because it exercises BibleUI current-document switch planning, Android category mismatch rejection, atomic auxiliary-document persistence, module-only selection, and category reload decisions without SWORD fixtures or app bootstrap.
 - `AndBibleTests+AppAndReader` reader chrome slice has moved to `BibleUITests` because it exercises BibleUI header, toolbar action, popup placement, overflow menu, Downloads routing, and keyboard shortcut construction without app delegate/bootstrap behavior.
+- `AndBibleTests+AppAndReader` module-browser slice has been split by owning module: Downloads filtering/sorting, failed-source cache merge, and startup-default installable-row guards moved to `BibleUITests`; Android metadata token decoding and catalog `InstallSize` byte preservation moved to `SwordKitTests`; the duplicate recommended-document refresh failure test was removed from the app-host bundle because equivalent SwordKit package coverage already exists in `SwordManagerTests`.
 - Batches for the genuinely BibleUI-behavior remainder: ReaderNavigation, Bookmarks/Strongs, Window/Settings, view-model parts of `+AppAndReader`.
 - **Split `+AppAndReader` deliberately**: `sceneConfiguration` test stays app-hosted; the `ContentView` legacy root-sidebar source scan is now a repo-standards `source-guards` check; the remaining package-owned tests should move to the lowest owning package target.
 - Run BibleUITests via `xcodebuild test -scheme BibleUITests -destination 'platform=iOS Simulator,...'` against the **committed package test scheme from Phase 0** (no app host).
